LayoutTests/inspector/console-format.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="console-tests.js"></script>
<script>

function doit()
{
    console.log('Message format number %i, %d and %f', 1, 2, 3.5);
    console.log('Message %s for %s', 'format', 'string');
    console.log('Object %o', {'foo' : 'bar' });

    var array = ["test", "test2"];
    array.foo = "bar";
    array[4] = "test4";
    console.log(array);

    dumpConsoleMessages();
}

</script>
</head>

<body onload="onload()">
<p>
Tests that console logging dumps proper messages.
</p>

</body>
</html>
