LayoutTests/inspector/console-format-expected.txt

CONSOLE MESSAGE: line 9: Message format number %i, %d and %f
CONSOLE MESSAGE: line 10: Message %s for %s
CONSOLE MESSAGE: line 11: Object %o
CONSOLE MESSAGE: line 16: test,test2,,,test4
Tests that console logging dumps proper messages.

console-format.html:9Message format number 1, 2 and 3.5
console-format.html:10Message format for string
console-format.html:11Object Object
console-format.html:16["test", "test2", undefined, undefined, "test4"]
